Use weighted average rather than k-nearest neighbor for downsampling

+++ src/apu/utils.rs
@@ -8,8 +8,8 @@ pub fn bounded_wrapping_add(original_value: u8, max_value: u8) -> u8 {
     new_value
 }
 
-pub fn as_dac_output(dac_input: u8) -> f32 {
-    (dac_input as f32 / 7.5) - 1.0
+pub fn as_dac_output(dac_input: f32) -> f32 {
+    (dac_input / 7.5) - 1.0
 }
 
 const LENGTH_ENABLED_INDEX: u8 = 6;
